Refactor: トランザクション関連のIFと実装を修正
