+++ .github/release/v0.18.0.md
@@ -3,6 +3,7 @@
 
 - uuid生成用の`pkg/uuid`を導入しました。
 - error生成・操作用の`pkg/xerrors`に更新しました。
+- トランザクション機能の管理場所を変更しました。
 
 ## 更新内容
 
@@ -11,6 +12,10 @@
   - これにより外部ライブラリの依存を減らし、ドメインなどのロジックで比較的安全にUUIDを運用できるようになりました。
 - `pkg/xerrors`
   - Readme.mdを更新し、`pkg/xerrors`の使用方法を明確にしました。
+- トランザクション管理機能を提供する`internal/usecase/tx_manager.go`を追加しました。
+  - トランザクションの開始、コミット、ロールバックを簡素化。
+  - `DoWithResult`関数を使用して、トランザクション内で値を取得可能。
+- トランザクション管理を`internal/usecase/tx_manager_test.go`に変更しました。
 
 ## 追加ライブラリ
