+++ internal/usecase/tx_manager.md
@@ -1,0 +1,31 @@
+# Tx Manager
+
+概要: このモジュールは、トランザクション管理を行うための機能を提供します。トランザクション内での操作を簡素化し、エラー処理を一元化します。
+
+## 役割
+
+- トランザクションの開始、コミット、ロールバックを管理します。
+- トランザクション内で値を取得するためのヘルパー関数の`DoWithResult`を提供します。
+
+## 必要度
+
+### 本番運用での必須度
+
+- 必須度: 本番運用で必須
+  - 理由: トランザクション管理は、データの整合性を保つために不可欠です。本番環境では特に重要です。
+
+### 開発/テスト運用での必須度
+
+- 必須度: 開発/テスト運用で必須
+  - 理由: 開発環境やテスト環境でトランザクションを使用することで、データの一貫性を確認しやすくなります。
+
+### 無効化した場合の影響
+
+- トランザクションが適切に管理されなくなり、データの整合性が失われる可能性があります。
+- エラー処理が複雑化し、コードの保守性が低下します。
+
+## 注意点
+
+- トランザクション内での操作は、必ず`Manager`インターフェースを通じて行ってください。
+- `DoWithResult`を使用する際は、ジェネリクスを正しく設定してください。
+- トランザクションのスコープを明確にし、不要なロックや競合を避けるように設計してください。
